Move CircularQueue and Array algorithms out of subfolders
