+++ firmware/shared/util/arrays.hpp
