+++ firmware/shared/util/circular_queue.hpp
